Local Multiplayer
Local multiplayer is now all done except for a win condition. It has the same result when the players collide or if one dies. That end game photo needs to be updated, but other than that, it is fully functionable local multiplayer. Player 1 is operated with WASD while Player 2 is the arrow keys. Would reccomend looking through the code. Since I made a different script (just an edited GameStage), the scripts such as Snake and SnakePart had to be edited in order to check what type of game (single or local multi) is being played. These checks are booleans that the stage passes when creating each object.